urban-hydrology-engine/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
geoalchemy2==0.15.1
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pulp==2.8.0
shapely==2.0.4
geojson==3.1.0
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
apscheduler==3.10.4
rasterio==1.3.9
numpy==1.26.4
